include: renamed auth_list.h -> auth_record.h
