+++ src/bpf/auth_check/auth_check_execve.bpf.c
@@ -3,7 +3,7 @@
 #include <bpf/bpf_core_read.h>
 #include <bpf/bpf_tracing.h>
 
-#include <auth_list.h>
+#include <auth_record.h>
 #include <auth_check.h>
 #include <k1_map.h>
 #include <k1_limits.h>
