+++ src/include/auth_record.h
