+++ src/k1cli/opt.c
@@ -6,7 +6,7 @@
 #include <verdict.h>
 #include <inttypes.h>
 
-#include <auth_list.h>
+#include <auth_record.h>
 #include <bpf_progs.skel.h>
 #include <common.h>
